+++ jest.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  preset: 'jest-expo',
+  testEnvironment: 'node',
+  verbose: true,
+};
