+++ package.json
@@ -4,8 +4,10 @@
     "start": "expo start",
     "android": "expo start --android",
     "ios": "expo start --ios",
-    "web": "expo start --web",
-    "eject": "expo eject"
+    "test": "jest",
+    "test:watch": "jest --watchAll",
+    "prettier:write": "npx prettier --write **/*.{js,ts,tsx,json}",
+    "lint": "tsc --noEmit && eslint --ext .js,.jsx,.ts,.tsx ./src/"
   },
   "dependencies": {
     "expo": "~42.0.0",
@@ -17,8 +19,13 @@
   },
   "devDependencies": {
     "@babel/core": "^7.9.0",
+    "@types/jest": "^26.0.24",
     "@types/react": "~16.9.35",
     "@types/react-native": "~0.63.2",
+    "eslint": "^7.30.0",
+    "jest": "^27.0.6",
+    "jest-expo": "^42.0.0",
+    "prettier": "^2.3.2",
     "typescript": "~4.0.0"
   },
   "private": true
