+++ src/__tests__/hello.test.ts
@@ -1,0 +1,5 @@
+describe('some tests', () => {
+  it('calculates 2+2=4', () => {
+    expect(2 + 2).toBe(4);
+  });
+});
